test: unify test cases
